[nggb] Start preparing next-gen GNU build scripts.
The new versions will be backward incompatible, so keep the existing
scripts intact. Users of existing build scripts are not affected. This
is to avoid having to continuously update documentation for multiple
platforms and distributions.

All modifications and replacements are in *.nextgen, which are copied
into their final location by activate.py. MD5 sums of legacy scripts are
included in *.nextgen, so that chagnes made to legacy scripts do not get
missed.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12916 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd