Fix hacks involving CDeviceBase [1/8]: ASIStage.
Eliminate the fake interface class ASIDeviceBase (from which the
concrete device classes don't actually derive, making the static_cast
from concrete device pointer to ASIDeviceBase* a dangerous hack).

Modify the ASIBase class to use MM::Core* (the call-back-to-core
interface) instead of CDeviceBase methods, so that it works without
having to have a pointer to the CDeviceBase interface of the concrete
device classes (which is what ASIDeviceBase* was used for).

Make ASIBase destructor virtual so that it actually gets called.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@13378 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd