Changes to Python export code proposed by Daferna.  Add DoubleVector to enable loadXYStageSequence()
loadXYStageSequence() requires a vector<double>

lines 203-209 needed to be moved down because SWIG %apply works indiscriminately and will match types inside the source code for std::vector causing failure to compile. It may be better in the future to use %apply and then %clear, or use a more specific name & type matching, but this works for now.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16421 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd