Dragonfly: Modify TIRF mode behaviour
If a mode is not selected, its associated value can't be modified. If
the user tries to modify its value, the change will be reverted.

In order to be compatible with presets, if the value associated to a
mode is modified when the mode is not selected, the value will be
remembered so that it can be set to that value when the mode is later
selected