Dragonfly: Improvement of the TIRF mode selection
When a mode is not selected, we now allow the user to change its
associated value unless Critical Angle is selected and Penetration is
the value being changed
However the value is only sent to the device when the user selects the
mode

Pros:
_ Previously, changing a value of a non selected mode was ignored at
first in the UI (the displayed value would be switched back to the
previous position) but the change was eventually accepted when
switching to that mode which felt awkward
_ Previously, loading a preset containing all TIRF related properties
would usually lead to only a part of the preset being visually loaded
since values associated to non selected TIRF modes would not be visibly
changed. As a result MicroManager would not display the preset's name
in place of the currently loaded preset possibly making the user
believe that it wasn't properly loaded. This is now working as intended
in most cases (see cons for the one case where it's not).

Cons:
_ If Critical Angle is the current mode and the user loads a preset
with HiLo selected, MicroManager will randomly not modify the displayed
value for Penetration (even though the value has been loaded properly
in the device adapter). This is only a glitch but may be confusing.
_ If the user modifies the value associated to a TIRF mode that is not
selected, the UI will update but the device will only be updated when
that mode is selected. This may be considered confusing to users.