Dragonfly: Improve disk speed monitoring
WARNING: do not use this commit with an actual device. Disk speed is
not read from the device.

Improvements to the disk speed:
_ the speed is saved to config file to be loaded on initialisation
(since the speed of the disk on initialisation of the real device is 0)
_ all monitoring properties are refreshed properly using both
Prop->Set() and GetCoreCallback()->OnPropertyChanged()
_ the speed of the disk is considered stable when it is fluctuating
less than 20RPM around the requested speed

The code currently uses a disk speed simulator. The next commit (which
will refactor some of the changes of this one) will switch back to the
actual device.