ILE: Add very low power property plus improvements
ClassIV or Key Interlock do not directly disconnect from the device.
It was causing crashes with a real device since it was happening in a
working thread and we didn't want to include mutexes in UI related
code. All classes were already checking for the state of the interlock
in every action methods so it wasn't required anymore anyway.

ActivateInterlock() is now called as soon as the pointer to the Reset
property is initialised if required. Some interlock may be active on
initialisation of the device so that ensures the property is read/write
as soon as possible.

Fire() and GetOpen() now properly return an error if there's one

Add some error reporting
Refactore interlock management in Lasers since it was a mess
Refactore a bit the main class