Nico: svn commit 16757 SerialManager: Improved performance on Windows. Pull Request #648 by Lukas-Lang: Make Serial port communication faster for USB based serial ports. See https://stackoverflow.com/q/48976397 Tested on a complicated microscope system and no stability problems found.  Device-Property manager appears to be respond much, much faster (as predicted by Lukas: going from ~70 to ~7ms per serial command).
This PR changes the timing settings for serial ports under Windows. Essentially,
 this speeds up USB based serial ports significantly - I have observed an improv
ement of up to 10x (from ~70ms to ~7ms per command) compared to the previous imp
lementation. This makes navigating using serial port based stages significantly
smoother.

The new timing settings effectively tell Windows to complete the read calls to t
he serial port immediately instead of waiting for a while. For a more detailed e
xplanation see this post on StackOverflow.

Since I do not have access to a computer with a true serial port, I could not quantify the impact on this type of port. However, my understanding is that it should not negatively affect performance.