nico svn r16801: VariLC: Pull Request #655 by nanthony21. This pull request improves the stability and versatility of the variLC.
Previously the adapter would cause a sudden crash of MM if a wavelength outside
of its tunable range was specified. Now it will log an error and prevent the cra
sh from occuring.

Rather than assuming that the tunable range of the device is 400-800 nm the adapter will now query this information from the device during initialization. Different models have different tunable ranges.