VarispecLCTF adapter by Nikc Anthony.
The existing VariLC adapter can be used to control the Varispec LCTF even though it wasn't originally written for that purpose and I have found that there are a number of bugs that cause MicroManager to crash. The new adapter fixes these bugs and adds a few extra functionalities such as property sequencing.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16891 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd