Core: Fix problem caused by differing interpretation of what constitutes a group between Core and UI
When a group has only a single property, the UI will show that single property 
and let the user manipulate its settings in the ConfigGroupPad.  However, the
OnConfigGroupChanged callback will notice that a new value (usually) does not 
match the group, hence signal that the group does not match any of its presets.
This only causes a problem in callbacks (caused by devices that use callbacks).
This change will now only issue the OnGroupChanged callback if the config group contains more than one property.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16959 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd