CoherentRemote: Remove line breaks from property names.
Property names are constructed by combing the laser name (obtained from the 
hardware) with identifiers.  The hardware apparently can return names 
that contain line breaks that were not removed.  This commit removes 
trailing white space from the laser name, thus removing line breaks 
from the middle of property names.  It is possible that some configurations 
will be broken because of this change, but this commit should allow 
the use of properties from this library in property groups even when 
the hardware insers line breaks.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17185 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd