PicardStage, PR #86 by Francois Marelli:
I slightly modified the PicardStage DeviceAdapter so that the CMMCore.home() function can be used to home the Z stage, with a behaviour identical to the XY stage.
I also changed the busy() function of the XY stage so that it is now considered busy when the stage is homing (thus having a behaviour consistent with the Z stage).

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17235 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd