Fix mapping of MMCore errors to Python exceptions
We are using SWIG's exception.i library (the language-neutral way to
raise exceptions;
http://www.swig.org/Doc4.0/SWIGDocumentation.html#Library_nn17), and the
correspondence to Python exceptions is defined here:
https://github.com/swig/swig/blob/90cdbee6a69d13b39d734083b9f91069533b0d7b/Lib/python/pyerrors.swg#L5-L46.
The Python exceptions are defined here, of course:
https://docs.python.org/3/library/exceptions.html.

Changed exception mapping to avoid SWIG_NullReferenceError
(undocumented?) and SWIG_SystemError (Python's SystemError is for
internal problems in the Python interpreter); use MemoryError only for
its intended meaning (out of memory); and use IOError for syntax errors
in configuration files.

MMERR_CameraBufferReadFailed is returned when a camera device adapter
returns a null pointer when the snapped image is requested. This could
be considered an IOError (if there was a device error during the
previous snap), but an equally likely cause is that there is no snapped
image (yet) to be returned (in which case RuntimeError is more
appropriate). Because most device errors will be handled by the default:
clause as RuntimeError anyway, it's probably simplest, from the user's
perspective, to do the same for MMERR_CameraBufferReadFailed.