Protect against serial port reopening failure.
With serial devices using FTDI USB-serial chips, reloading a config file
would fail stochastically. (Probably this would not be seen with a
config file containing more than 1 or 2 devices.)

Win32 API call tracing showed that the port and related resources are
correctly closed, so added an extra sleep after joining the async
thread, which eliminated the problem in my testing. Sleeps are 2x the
minimum I found needed in testing, to account for system variability.