Use Boost 1.77.0 for VS2019 build
It would have been nice to keep the old Boost version and just build it
for VS2019, but that does not work because automatic linking (via
pragmas in the Boost headers) fails with Boost 1.55.0 due to not knowing
about MSVC 13+.

So update to 1.77.0 (the latest). This introduces a few errors due to
incompatibility (in SerialManager and DEClientLib).