Update OpenCV for FakeCamera
Point to OpenCV 2.4.13.6 built with VS2019. FakeCamera relies on the
OpenCV DLLs installed by OpenCVgrabber's build.xml.

This updates to the equivalent of the previous (VS2010) OpenCV 2.4.8
build, but I suspect the OpenCV build used here may not be the most
useful for FakeCamera, because it does not include the optional image
I/O libraries that OpenCV can make use of. It may be appropriate to
revisit this after we finish moving to VS2019.