Dragonfly: Fix TIRF mode selection change
TIRF mode associated value (Critical angle offset,
HiLo oblique angle or Penetration) is buffered if the current TIRF
mode is not the correct one.
In this case, when selecting the correct TIRF mode, the buffered
value has to be used to update the device.
This was not working properly and is now fixed.