Require explicit construction of MMTime
Forbid implicit conversion from double (so that, for example, arithmetic
and comparison operators won't treat double as microseconds but instead
require an explicit MMTime object). This is safer because there is a
lower risk of getting the units wrong.

While this change requires device code to be updated, it does not affect
the binary interface and therefore does not require the device interface
version to be incremented.

Changes to device adapters all belong to these categories:

- Comparison of an MMTime with a number -> replace number with MMTime
  using appropriate unit (`1000.0 * x` -> `MM::MMTime::fromMs(x)`).

- In a subset of the previous case, remove a redundant if statement and
  return a bool directly.

- Replace assignment of 0 with assignment of MM::MMTime{}.

- Replace `return 0` with `return {}` in functions returning MMTime.

- Subtraction of a number from MMTime -> replace with MMTime using
  appropriate unit (fromSeconds())

All of the above changes preserve behavior exactly. But a few obvious
bugs were fixed:

- K8601/K8601.cpp:539
  Fix `1000.0 < GetDelayMs()` which should have been `1000.0 *
  GetDelayMs()`; use `MM::MMTime::fromMs(GetDelayMs())`.
  Compare K8055 and others that use the same pattern.

- Yokogawa/CSUX.cpp:330
  Comment said '33 msec' but use of raw number meant it was interpreted
  as 33 us. Fixed to use 33 ms as this was almost certainly what was
  intended.

The following looks like a potential bug, but I'm leaving the behavior
unchanged in the absence of a means to test:

- PeCon2000/PeCon2000.cpp:378
  Comparison with `double delta` interprets `delta` as microseconds, but
  usage of `is_older_than` under variable name `is_older_than_a_minute`
  suggests `delta` (= 60.0) was meant to be in seconds.
  I cannot tell if the behavior is correct or the name, so I'm opting to
  preserve current behavior.