Require SWIG 2.x or 3.x to build MMCoreJ
Building with SWIG 4+ results in errors, possibly later in the build (in
MMStudio, because SWIG 4 changed some of the wrapped types). Until now,
this was confusing and has resulted in multiple issues filed.

Until we have a proper fix for SWIG 4, let's make it fail immediately
with the correct error message.