Retire FixSnprintf.h
This was only needed to compile with Visual Studio 2010.
It is no longer necessary with modern MSVC (doesn't do anything).

Just in case, checked that the return value of snprintf() is not used
(since it deffers from that of _snprintf()). It was only used by
WOSM_rev.cpp, which is a recently added file (never built with old
MSVC).