Zaber: Skip build when ZML not found
Closes #368.

Without this, 3rdpartypublic is required for the Mac/Linux build. We
want to keep it so that only 3rdpartypublic/classext is required on
those platforms.

This change also removes the special rpath flags. In the official macOS
build, dependency paths are handled by a spearate script. On Linux, the
only currently supported way is to build on the system where MM will be
run, so (I hope) it is not necessary to use a special flag.