Fixed UNI-583 VimbaX shared libraries must be copied to "/usr/local/bin"
The Allied Vision Camera Adapter was trying to load the VimbaX shared libraries from the fixed path "/usr/local/bin". This was not good, because "/usr/local/bin" is on linux a location for applications and not for libraries. Instead, now the standard linux dynamic library loading mechanism is used, which tries to automatically resolve the path of the shared library by using the system paths, ld.so.conf and the environment variable LD_LIBRARY_PATH.
Adjusted the licence headers, because the micro manager adapter development was not started in 2012.