Fixed UNI-591 main window exposure setting not working
The legacy cameras use the feature "ExposureTimeAbs" instead of "ExposureTime" for setting the expsure time. The adapter was using the "ExposureTime" feature for the main window exposure setting. If the feature "ExposureTime" is available, it is used for the exposure setting. But if it is not available now "ExposureTimeAbs" is used for the exposure setting. The value of the exposure setting now also corrected to milliseconds.