Fixed float property change always called
Inside onProperty for every float property the OnPropertyChanged callback was always called even when the value has not changed, because for the comparison  the float were converted to strings with a different count of decimal places.
On stream start and stop now OnPropertiesChanged is called to update the ReadOnly status of all properties in the gui.