MMCore: Revert strict initialization checks
When an operation that requires an initialized device is attempted on an
uninitalized device, log a warning message instead of throwing an
exception.

It turns out that the Hardware Configuration Wizard currently relies
(mostly incorrectly) on some operations that are not actually correct,
but fixing it is complicated enough that we need to revert the exception
throwing for now.

In addition, MMCore itself needs some fixes so that functions like
waitForSystem() do not cause issues in the presence of uninitialized
devices. This may also apply to other bulk actions such as
unloadAllDevices() and the system state cache.

The goal is still to enable exceptions for these checks, but only once
we've had a chance to fix these issues.

This does not revert the checks against multiple initialization attempts
and against setting a pre-init property after initialization. These are
not known to cause any issues so far.

Also add the name of the operation to the logged warning message to
assist with the necessary fixes. The log messages have log level `WRN`,
which should stand out since we have never used it before. This is to
prevent people from thinking that the "normal" operation of the Hardware
Configuration Wizard is causing errors.

Bump MMCore version to 10.6.0.