Remove MMDevice/MMCore unit tests from Automake
We do not run these tests from CI using Automake currently, and there is
not much of an advantage in doing so because we now have them run using
the experimental Meson build for MMDevice and MMCore.

So remove from the Automake build so that these tests can be evolved
(including switching the testing framework away from GoogleTest) without
having to update the Automake build.

For the couple of device adapters that have unit tests, those can still
be run from Automake.