Replace MODULE_EXPORTS with MMDEVICE_CLIENT_BUILD
The new macro has (roughly) the opposite meaning as the (intent of)
MODULE_EXPORTS: it should be defined when building MMCore and undefined
when building device adapters.

The source is now _capable_ of excluding the module interface functions
when building MMCore (and the MMCore Meson build does so).

Symbol visibility is set for GCC/Clang so that -fvisibility=hidden can
be used when building device adapters (but this is not done for now).

No changes to device adapters is required.