MMDevice/MMCore[J]: Remove unused macro defs
Remove preprocessor definitions from MMDevice/MMCore/MMCoreJ_wrap
.vcxproj files

This is to maintain the cleanup done in #415 (specifically,
4536cc099caa4793fadacbea3dcc694de679f014) -- to prevent the vcxproj
build from going out of sync with the experimental Meson build.

- `_DEBUG` - the compiler defines this when using the debug runtime; redundant
- `_LIB` - added by VS by default for static libs; we don't use it
- `_USRDLL` - added by VS by default for DLLs; we don't use it
- `_WINDOWS` - added by VS by default; we don't use it
- `WIN32` - added by VS by default; we prefer `_WIN32`, defined by the compiler

Also delete the single use of `_DEBUG` in `CMMCore::getVersionInfo()`.
It is better not to produce a different version string in a debug build,
because applications may parse this string (for example, a pymmcore test
does that).

All the other macros removed here are not currently used in these
projects.