MMDevice: Remove need for _CRT_SECURE_NO_WARNINGS
That is, remove use of strcpy(), strncpy(), and getenv().

Replace str[n]cpy() with snprintf(). Add tests in the case of
CDeviceUtils::CopyLimitedString() (previous implementation had a bug
where it didn't null-terminate if the source string was exactly
MM::MaxStrLength - 1 chars long).

The use of getenv() was dead code, so remove entirely.