Convert possible CP-1253 to UTF-8
This file already contained ASCII '?' characters, suggesting a previous
encoding conversion had already lost some characters. (This seems to
have happened before the file was first committed (2012) to this
repository or its predecessors.)

The character in question here is 8-bit F8, which in CP-1252 (or latin1)
would be 'ø'. Maybe that is what it was, but in CP-1253 (iso-8859-7,
Latin/Greek) it is 'ψ', which seems more likely here.

In any case, we are not losing any information by this conversion.