Upgrade micromanager.sln to VS2022
This does not touch the projects (still VS2019) but just updates the
header in micromanager.sln.

Since most people are probably now using VS2022, this will avoid having
to manually exclude changes when e.g. adding a project.

I don't think this prevents the solution from opening in VS2019 (not
sure, but we probably don't care).