Prevent boost::bind warning on _1, _2, ...
Prevent the following warning:

.../boost/bind.hpp:36:1: warning: The practice of declaring the Bind placeholders (_1, _2, ...) in the global namespace is deprecated. Please use <boost/bind/bind.hpp> + using namespace boost::placeholders, or define BOOST_BIND_GLOBAL_PLACEHOLDERS to retain the current behavior.

New code should use lambdas rather than boost::bind or std::bind, but
that transformation is not quite trivial, so keep boost::bind so as not
to risk breaking the code. Instead follow the warning message's advice
and use `using namespace boost::placeholders` locally.

In the case of SequenceTester, however, lambdas are simpler so use them.