MMCore: Unset device role when unloading
Regression test (for camera only, for now) added.

Before this change, an exception would be thrown when trying to unload
the current camera, etc., because the property no longer had the current
device's label as an allowed value, yet the weak_ptrs to the current
devices remained lockable due to shared_ptr(s) held to the unloaded
device.

Proactively removing the role assignments for the device, as done here,
should work regardless of who happens to hold a shared_ptr to the
device.