Meson: Allow MMCore to fetch mmdevice via wrap
This allows mmcore (from the mirror repo) to build automatically.

But keep the manual copy for CI and justfile (to use correct version of
MMDevice).

The mmdevice.wrap just fetches the latest HEAD. Once MMCore is no longer
in mmCoreAndDevices it can point to a specific mmdevice commit, but that
would only be confusing if done now when official MMDevice is a
cohabitant of mmCoreAndDevices. (The mmdevice commit would never be up
to date!)