Meson: MMDevice: default to hidden visibility
MMDevice is either built into device adapters or (usually, via MMCore)
into Python or Java extensions, which are all DLLs/shared objects. None
of these should expose the MMDevice classes and functions (except for
the ones marked MODULE_API). (On Windows, these are automatically hidden
because they are not exported.)