Add PropertyShutter utility device
Implements a shutter that can control any device property, not just digital-to-analog outputs (as in DAShutter with Arduino and NIDAQs).

Enables using sliders or other properties as shutters (e.g., a halogen lamp with only a voltage slider and no explicit on/off property).

Useful for MDA imaging where such devices need to act as shutters.

Verified on a Zeiss microscope using the ZeissCAN29 device adapter.

The following property can be set:
- Target Device & Target Property: Which property should the PropertyShutter change
- Open & Closed Value: To which state the property should be changed when the shutter is opened / closed
- Shutter Delay: An additional delay after opening / closing the shutter can be set here.
- Invert Logic: Should the open / closed values be inverted for opening / closing the shutter.
- State: Usefull for debugging, instead of opening / closing the shutter through the GUI, you can set it here as a property