Fix cameras that were missing just Busy()
Restore compilation of these cameras without changing behavior.

Rather than inheriting from CLegacyCameraBase, simply supply a Busy()
that returns false, which is the same behavior as with
CLegacyCameraBase::Busy(), given that no other legacy functions are used
in these cameras.

Left a comment on Multi Camera's Busy(), because theoretically it should
forward the physical cameras' Busy() results. However, it is out of
scope for this commit to change behavior (and Busy() on cameras
traditionally does not get used).