CCameraBase -> CLegacyCameraBase where necessary
Made sure that all occurrences of CCameraBase is replaced in these
cameras, so that behavior should match previous.

Some of these cameras only make minor use of CLegacyCameraBase member
functions, and can be made to work with the new CCameraBase with little
work, but doing the simplest transformation for now.