Remove GetNominalPixelSizeUm() from cameras
It appears that, although this function has not been in the MM::Camera
interface since at least 2007 (if ever), a lot of cameras have included
it -- many by copy-pasting.

A couple of cameras actually had code to get the camera's sensor pixel
size, but it is not useful at this time to keep those because (aside
from never having been tested) the majority of cameras just returned
1.0, leaving no practical way for an application to make use of the
value. (If we want this functionality, we should define a standard but
optional property and require cameras to only provide the value if
known.)

So remove the dead code.