Remove redundant base class default initializers
C++ always calls a base class constructor, and there is no need to make
an explicit call when using the default constructor of the base class.

The fact that these show up in cameras, stages, and XY stages (but not
other device types) suggests all this is the result of copy-paste.

Since I recently removed a bunch of these (for cameras that needed to
switch from CCameraBase to CLegacyCameraBase), I'm taking the
opportunity to eradicate them.