Use globally unique handle assigned by Core
Using two different integers ("core" and "device" handles) to represent
the same dataset is unnecessary.

Instead, let the Core assign a globally unique dataset handle, and
eliminate handle assignment from devices entirely. This also avoids
copy-pasting the handle assignment logic every time writing a new
storage device.

In the Core, instead of mapping handles to device labels
(`std::string`), map to `std::weak_ptr<StorageInstance>`. This is
important, because otherwise there could be ABA problems: the storage
device with the same label may not be the same device, or it may have
been reloaded, invalidating all handles. (That could also be prevented
by making sure to remove handles when devices are unloaded, but it's
better to eliminate even the possibility of such an error.)