AMF: Fix vcxproj header/source grouping
AMF_LSP_Pump.h was listed as a source file. Compiling it results in
AMF_LSP_Pump.obj, which conflicts with the output of compiling
AMF_LSP_Pump.cpp (if the compiler tries to write at the same time). This
caused sporadic build failures. At least, that's what I think was
happening.