MMDevice Meson: Disable tests for device adapters
If we ever build more than a few device adapters at the same time
(without using a parent Meson project), each build will build MMDevice.
This is fine because MMDevice builds fast, but not great if MMDevice
tests are enabled because then Catch2 is downloaded and built (unless
found on the system).

However, we probably want the device adapter's own tests (if any) to be
enabled by default.

So disable MMDevice tests by default when MMDevice is being built as a
subproject.

But don't disable tests if MMDevice is being built as a subproject of
MMCore. This is important because any tests specific to the
client_interface build might never get exercised otherwise. And we don't
have many MMCores.

When building a device adapter, it will still be possible to enable
MMDevice tests with -Dmmdevice:tests=enabled.